backend/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
